feat(fuzz,membrane): add phase 2 fuzz targets, evidence ledger, and SLO contracts
Phase 2 Fuzzing:
- Add fuzz_malloc target (117 lines): structure-aware malloc fuzz harness
  covering allocator edge cases (calloc overflow, realloc cycles)
- Add fuzz_runtime_math target (129 lines): fuzz harness for runtime
  math kernel decision testing (determinism, risk bounds, action invariants)
- Add seed corpora for malloc and runtime_math targets
- Add fuzz phase 2 target architecture specification (780 lines JSON)
- Add generate_fuzz_phase2_targets.py for harness architecture generation
- Add check_fuzz_phase2_targets.sh validation script
- Add fuzz phase 2 targets regression test
- Update CI pipeline and nightly fuzz script for new targets

Evidence Ledger (new module):
- Append-only ledger for TSM (Transparent Safety Membrane) evidence
  collection with JSONL export and configurable redaction policies
- Supports structured event recording for safety audit trails

SLO Contracts (slo.yaml):
- Define SLO targets for data and decision planes with per-pipeline-stage
  budgets and reliability guarantees

Membrane Improvements:
- EBR, heal, ptr_validator, and lib updates supporting evidence
  collection integration
- Add membrane examples directory

Beads:
- Close bd-1om (golden snapshot regression gate)

+++ crates/frankenlibc-membrane/src/ptr_validator.rs
@@ -760,7 +760,7 @@ impl ValidationPipeline {
                             state,
                             cv.user_base,
                             remaining,
-                            u64::from(cv.generation),
+                            cv.generation,
                         );
                         self.runtime_math.note_check_order_outcome(
                             mode,
